Remove Me::Store2; use store2 gem instead
